Refactor: Remove GroupAnchorOffsetOverlay component and related files; reorganize ToolbarOverlay for better structure
